refactor: 스타일도 컴포넌트 역할별로 분리
