+++ src/app/home/components/ExperimentPostContainer/FilterContainer/AreaFilter/AreaFilter.css.ts
@@ -1,5 +1,4 @@
 import { style } from '@vanilla-extract/css';
-import { recipe } from '@vanilla-extract/recipes';
 
 import { colors } from '@/styles/colors';
 import { fonts } from '@/styles/fonts.css';
@@ -43,64 +42,16 @@ export const contentWrapper = style({
   display: 'flex',
 });
 
-// 지역 목록 컨테이너
-export const areaListContainer = style({
-  flex: '0.5',
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '0.4rem',
-  padding: '1.2rem 0.8rem',
-  backgroundColor: colors.field02,
-  borderRadius: '1.2rem',
-  height: '28rem',
-  overflow: 'scroll',
-});
-
 // 기본 텍스트 스타일
 export const areaName = style({
   ...fonts.label.large.M14,
   color: colors.text06,
 });
 
-// 선택된 지역 텍스트 스타일
-export const selectedRegionName = style({
-  color: colors.textPrimary,
-});
-
 // 숫자 표시 스타일
 export const areaCount = style({
   ...fonts.label.medium.R13,
   color: colors.text03,
-});
-
-// 지역 버튼을 recipe로 정의 (selected 여부에 따라 스타일이 변경됨)
-export const areaButtonRecipe = recipe({
-  base: {
-    display: 'flex',
-    gap: '0.4rem',
-    alignItems: 'center',
-    padding: '0.6rem 1.2rem',
-    borderRadius: '1.2rem',
-  },
-  variants: {
-    selected: {
-      true: {
-        backgroundColor: colors.primaryTinted,
-        outline: `0.1rem solid ${colors.lineTinted}`,
-        fontWeight: 'bold',
-      },
-      false: {
-        selectors: {
-          '&:hover': {
-            backgroundColor: colors.field03,
-          },
-        },
-      },
-    },
-  },
-  defaultVariants: {
-    selected: false,
-  },
 });
 
 // 세로 구분선
@@ -119,88 +70,6 @@ export const verticalLine = style({
     },
     '&:last-child::after': {
       display: 'none',
-    },
-  },
-});
-
-// // 서브 지역 리스트 컨테이너
-// export const subAreaListContainer = style({
-//   flex: 1,
-//   display: 'flex',
-//   flexDirection: 'column',
-//   gap: '0.4rem',
-//   height: '28rem',
-//   overflow: 'scroll',
-//   padding: '1.2rem 0.8rem',
-//   borderRadius: '1.2rem',
-// });
-
-// // 서브 지역 항목 (label) 기본 스타일
-// export const subAreaItem = style({
-//   display: 'flex',
-//   alignItems: 'center',
-//   justifyContent: 'space-between',
-//   gap: '0.4rem',
-//   height: '100%',
-//   padding: '0.6rem 0.8rem',
-//   borderRadius: '1.2rem',
-// });
-
-// // 선택된 서브 지역 항목 스타일
-// export const selectedSubAreaLabel = style({
-//   outline: `0.1rem solid ${colors.lineTinted}`,
-//   backgroundColor: colors.primaryTinted,
-// });
-
-// // 체크박스 숨김 처리
-// export const checkbox = style({
-//   position: 'absolute',
-//   opacity: 0,
-//   pointerEvents: 'none',
-// });
-
-// // 서브 지역 정보 컨테이너 (opacity를 CSS 변수로 제어)
-// export const subAreaInfo = style({
-//   display: 'flex',
-//   gap: '0.4rem',
-//   alignItems: 'center',
-//   opacity: areaOpacity,
-// });
-
-// // placeholder (지역 미선택 시)
-// export const placeholderArea = style({
-//   ...fonts.label.large.R14,
-//   color: colors.text03,
-//   display: 'flex',
-//   justifyContent: 'center',
-//   alignItems: 'center',
-//   height: '100%',
-// });
-
-// 버튼 스타일을 recipe로 재사용 (중복되는 padding, borderRadius 등)
-export const buttonRecipe = recipe({
-  base: {
-    ...fonts.label.large.SB14,
-    padding: '0.6rem 1.4rem',
-    border: 'none',
-    borderRadius: '1.2rem',
-  },
-  variants: {
-    type: {
-      reset: {
-        backgroundColor: colors.field03,
-        color: colors.text06,
-      },
-      save: {
-        backgroundColor: colors.primaryMint,
-        color: colors.text01,
-        selectors: {
-          '&:disabled': {
-            opacity: 0.6,
-            cursor: 'not-allowed',
-          },
-        },
-      },
     },
   },
 });
