+++ src/app/home/components/ExperimentPostContainer/FilterContainer/AreaFilter/components/AreaContainer/AreaContainer.css.ts
@@ -1,7 +1,5 @@
 import { createVar, style } from '@vanilla-extract/css';
 
-import { areaOpacity } from '../../AreaFilter.css';
-
 import { colors } from '@/styles/colors';
 import { fonts } from '@/styles/fonts.css';
 
@@ -19,7 +17,6 @@ export const subAreaListContainer = style({
   borderRadius: '1.2rem',
 });
 
-// 서브 지역 항목 (label) 기본 스타일
 export const subAreaItem = style({
   display: 'flex',
   alignItems: 'center',
@@ -30,20 +27,17 @@ export const subAreaItem = style({
   borderRadius: '1.2rem',
 });
 
-// 선택된 서브 지역 항목 스타일
 export const selectedSubAreaLabel = style({
   outline: `0.1rem solid ${colors.lineTinted}`,
   backgroundColor: colors.primaryTinted,
 });
 
-// 체크박스 숨김 처리
 export const checkbox = style({
   position: 'absolute',
   opacity: 0,
   pointerEvents: 'none',
 });
 
-// 서브 지역 정보 컨테이너 (opacity를 CSS 변수로 제어)
 export const subAreaInfo = style({
   display: 'flex',
   gap: '0.4rem',
@@ -51,7 +45,10 @@ export const subAreaInfo = style({
   opacity: areaOpacity,
 });
 
-// placeholder (지역 미선택 시)
+export const selectedRegionName = style({
+  color: colors.textPrimary,
+});
+
 export const placeholderArea = style({
   ...fonts.label.large.R14,
   color: colors.text03,
